+++ SourceCode/Services/Extensions/ICollectionExtensions.cs
@@ -1,0 +1,6 @@
+﻿namespace ModulesRegistry.Services.Extensions;
+public static class ICollectionExtensions
+{
+    public static bool IsNullOrEmpty<T>(this ICollection<T>? collection ) =>
+        collection is null || collection.Count == 0;
+}